requirements.txt
numpy==2.4.2
pandas==3.0.1
python-dateutil==2.9.0.post0
six==1.17.0
pdfplumber>=0.11
pdf2image>=1.17
pytesseract>=0.3.13
Pillow>=10.0
sentence-transformers>=2.7
requests>=2.31
